requirements.txt
flask
requests
trafilatura
PyPDF2
python-dotenv
transformers
torch
sentencepiece
sumy
nltk
